Update README.md
Implement a system that keeps track of students at a university. The system should keep personal details, the modules the students are studying and the marks achieved in each module. The system must be able to provide a series of reports, for example displaying the students' details and their marks, calculating the average mark, and displaying their result (pass, fail, distinction for example) based on a particular formula.